Continuation line character typo error
For section 4, Configure docker and Kubernetes prerequisites on the machine, the forward slash character is used at the end of each line as the continuation character. It should be a backslash character (or in this case could contain no continuation characters at all due the && characters).  

As it stands it won't deploy all the correct components (kubeadm for example).