Comment added that the run command is currently supported only for Python 3 notebooks.
Suggested update. The run command as of the current azdata version (15.0.1900) works for Python 3 notebooks, but not for other kernels such as SQL, Spark, Scala, R. Included a note to mention this. Not sure if totally necessary as will likely change in the future as other kernels are added. Included for awareness at the moment.